refactor: 지도 기능을 Feature Slice 구조로 재구성
- map feature를 완전한 Feature Slice 구조로 재조정
- FAB 버튼을 commons에서 map feature 내부로 이동
- WebView 관련 코드를 map-view 컴포넌트로 분리
- hooks, components, types 계층 구조 명확화
- utils, commons/layout/provider 폴더 추가
- 폴더 구조 문서화 (folder-structure.md)
