--- commons/components/fab-button/prompts/prompt.101.ui.txt
@@ -1,35 +1,0 @@
-아래의 조건을 모두 적용하여, 아래의 요구사항을 모두 구현할 것.
-구현 결과를 체크리스트로 반환할 것.
-
-==============================================
-
-조건-커서룰) 아래의 커서룰을 적용하여 작업하고, 이 작업이 끝나면 해당 rules 적용 결과를 체크리스트로 반환할 것.
-            - @01-common.mdc
-            - @02-wireframe.mdc
-            - @03-ui.mdc
-
-==============================================
-
-조건-피그마) 활용할MCP: Figma Desktop MCP
-조건-피그마) 메인 FAB 디자인 링크: https://www.figma.com/design/k7IWFISJsHIQ4g6FoAZqup/%ED%83%80%EC%9E%84%EC%BA%A1%EC%8A%90---%EC%9D%B4%EC%8A%A4%ED%84%B0%EC%97%90%EA%B7%B8?node-id=136-2309&t=gQWt4vOjwOGVJGCI-4
-조건-피그마) 서브 버튼 및 확장 상태 링크: https://www.figma.com/design/k7IWFISJsHIQ4g6FoAZqup/%ED%83%80%EC%9E%84%EC%BA%A1%EC%8A%90---%EC%9D%B4%EC%8A%A4%ED%84%B0%EC%97%90%EA%B7%B8?node-id=204-91&t=gQWt4vOjwOGVJGCI-4
-
-조건-파일경로) 구현될 TSX 파일경로: FE/commons/components/fab-button/index.tsx
-조건-파일경로) 구현될 CSS 파일경로: FE/commons/components/fab-button/styles.ts
-
-==============================================
-
-핵심요구사항) MCP를 연동하여 피그마의 디자인을 그대로 소스코드로 구현할 것.
-
-핵심요구사항) fab버튼이 눌렸을 때에는 "+"인 상태일 것.
-
-핵심요구사항) 눌렸을때, 사이즈 변동은 없을 것.
-
-핵심요구사항) 버튼 컴포넌트는 다음과 정의될 것.
-  - 기본 상태: 우측 하단에 노란색 원형 버튼 (+ 아이콘)
-  - 확장 상태:
-    - 배경에 반투명한 검은색 오버레이(Dimmed Layer)가 깔림.
-    - 메인 버튼은 X 아이콘으로 변경됨 (닫기 상태).
-    - 메인 버튼 위로 두 개의 서브 버튼이 수직으로 정렬됨.
-      - 위쪽: 이스터에그 (텍스트 라벨 + 아이콘 버튼)
-      - 아래쪽: 타임캡슐 (텍스트 라벨 + 아이콘 버튼)
