+++ commons/layout/provider/README.md
@@ -1,0 +1,15 @@
+# Layout Providers
+
+앱 전반에 걸쳐 사용되는 Provider들을 관리합니다.
+
+## 예시
+
+- ThemeProvider
+- AuthProvider
+- NavigationProvider
+- SafeAreaProvider
+
+## 원칙
+
+- Layout과 분리하여 관리
+- 전역 상태나 컨텍스트 제공
