+++ components/map/components/fab-btn/types.ts
@@ -1,0 +1,10 @@
+/**
+ * FAB Button Types
+ * Version: 1.0.0
+ * Created: 2025-12-17
+ */
+
+export interface FabButtonProps {
+  onEasterEggPress?: () => void;
+  onTimeCapsulePress?: () => void;
+}
