+++ components/map/components/map-view/hooks/useMapGestures.ts
@@ -1,0 +1,15 @@
+/**
+ * Map Gestures Hook
+ * Version: 1.0.0
+ * Created: 2025-12-17
+ *
+ * [UI Logic] 지도 줌/패닝 제어 로직
+ */
+
+export const useMapGestures = () => {
+  // TODO: 제스처 로직 구현 (필요시)
+
+  return {
+    // 제스처 관련 핸들러들
+  };
+};
