+++ utils/README.md
@@ -1,0 +1,16 @@
+# Utils
+
+[Pure Functions] 비즈니스 로직 없는 순수 함수
+
+## 예시
+
+- 날짜 포맷팅
+- 문자열 변환
+- 숫자 계산
+- 유효성 검사
+
+## 원칙
+
+- 앱의 도메인 로직을 알지 못함
+- 입력에 대해 항상 같은 출력
+- 사이드 이펙트 없음
